requirements.txt
python-telegram-bot==20.7
solana==0.30.2
solders==0.18.0
aiohttp==3.9.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
python-dotenv==1.0.0
numpy==1.26.2
cachetools==5.3.2
base58==2.1.1